Revert "Refactor project structure by removing main.py, adding config.py for environment variables, and updating database connection in database.py"
This reverts commit c2815ad34e5584539b664faaa02fa5e32afae83b.
