+++ main.py
@@ -1,0 +1,6 @@
+from fastapi import FastAPI
+
+
+app = FastAPI()
+
+
